+++ lib/flutter_esc_pos_utils.dart
@@ -6,4 +6,5 @@ export 'src/enums.dart';
 export 'src/pos_column.dart';
 export 'src/pos_styles.dart';
 export 'src/qrcode.dart';
+export 'src/pdf417.dart';
 export 'src/generator.dart';